+++ CHANGELOG.rst
@@ -1,11 +1,28 @@
 CHANGELOG
 =========
 
-6.0.0a15 (unreleased)
+6.0.0a16 (unreleased)
 ---------------------
 
 - Changes in ICatalogUtility, DefaultSearchUtility and @search endpoints
   [masipcat]
+
+- Update react-gmi v 0.4.0
+  [jordic]
+
+- Fix more antipatterns [lferran]
+
+- Fix integer query param validation [lferran]
+
+
+6.0.0a15 (2020-03-02)
+---------------------
+
+- Handle http.disconnect (and other types of messages) while reading the request body
+  [masipcat]
+
+- Be able to have async schema invariants
+  [vangheem]
 
 - Provide better validation for json schema field
   [vangheem]